Improve downloadHandler
